+++ megamek/src/megamek/common/MechSummaryComparator.java
@@ -1,14 +1,14 @@
 /*
  * MechSummaryComparator.java - Copyright (C) 2002 Josh Yockey
- * 
- *  This program is free software; you can redistribute it and/or modify it 
- *  under the terms of the GNU General Public License as published by the Free 
- *  Software Foundation; either version 2 of the License, or (at your option) 
+ *
+ *  This program is free software; you can redistribute it and/or modify it
+ *  under the terms of the GNU General Public License as published by the Free
+ *  Software Foundation; either version 2 of the License, or (at your option)
  *  any later version.
- * 
- *  This program is distributed in the hope that it will be useful, but 
- *  WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY 
- *  or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License 
+ *
+ *  This program is distributed in the hope that it will be useful, but
+ *  WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ *  or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
  *  for more details.
  */
 
@@ -52,7 +52,7 @@ public class MechSummaryComparator implements Comparator<MechSummary> {
         }
     }
 
-    private int numCompare(int n1, int n2) {
+    private int numCompare(float n1, float n2) {
         if (n1 == n2) {
             return 0;
         } else if (n1 > n2) {
